+++ internal/default/infrastructure/repository/user_impl.go
@@ -11,8 +11,8 @@ import (
 	"github.com/abyssparanoia/rapid-go/internal/default/domain/model"
 	"github.com/abyssparanoia/rapid-go/internal/default/domain/repository"
 	"github.com/abyssparanoia/rapid-go/internal/default/infrastructure/entity"
+	"github.com/abyssparanoia/rapid-go/internal/pkg/error/httperror"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/gluesqlboiler"
-	"github.com/abyssparanoia/rapid-go/internal/pkg/httperror"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/log"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
