+++ internal/pkg/error/grpcerror/model.go
