+++ internal/pkg/error/httperror/model.go
